make the models threadsafe
with this patch the unit test seems to work and I successfully
ran a complete kdevelop session through endoscope without a crash

Turns out, we cannot make a model threadsafe. While it itself
might be properly mutex protected, as soon as we use proxymodels
things get ugly: the {begin,end}{InsertRemove}Rows() calls trigger
signals which must be delivered directly to the proxy, otherwise
its internal state may be messed up and asserts start flying

So the solution: only call these methods in the main thread and
directly invalidate data on remove when called from a background thread.

Furthermore we introduce a mutex to the Probe, which gets locked
properly in objectAdded/Removed to synchrounize these events
which make the new isValidObject API behave sanely when called in
the proper places