extend unit test and test in-ctor cases
i.e. we now create instances of the new TestObject which in turn
creates a child QObject in it's ctor and also connects+disconnects inside the ctor

this way we get coverage of the related part in the probe and
indeed the threading test fails now as expected. more to fix