improve stability of gammaray when run on multi threaded applications
the probe now has a lock which can be accessed from the outside.
when locked you can be sure that valid objects (as indicated by
Probe::isValidObject()) stay alive and can be accessed.

furthermore we remove the last usages of QPointer as it's not threadsafe.

the unit tests now pass for me, but there is still a qWarning emitted
in QSortFilterProxyModel from time to time. Debugging showed that
the internal mapping gets cleared in _q_sourceDataChanged as called
by KRecursiveFilterProxyModel. I have yet to find out why this call
is in place and why this wipes the mapping