use our own queue which we can properly cleanup
the Qt::QueuedConnection queue could fire for already deleted
objects and hence confuse our models and assumptions, esp.
we must get the signal for it's parent before any child.
this used to fail if an object got created with the same
pointer adress of an object that was already deleted but
it's signal was not yet received