Add support for multiple inheritance to the meta object system.
The tricky part is casting void pointers to the correct base class, but
seems to work at least for QGraphicsObject (instances of which now also
show all QGraphicsItem properties in addition to the QObject ones).

It's limited to at most three base classes right now, to avoid introducing
a Boost MPL dependency, but that should be plenty for now and can always
be extended.