Fix passing of QApplication command line arguments during launching.
We were creating the QApplication object too early (and thus also when
only using the command line mode), which eats all arguments it can handle
(eg. -style or -graphicssystem), even if they were actually supposed to
go to the child process.
Now we only create it when showing the launcher window, but for that we
of course need to make the built-in arguments pass through our parser too.