Crash safety in the widgetinspector
We always assume there is a valid overlay widget instance, but in case
the target application is closed *before* GammaRay, the overlay widget
is deleted as well within ~QObject() (because it is a child of some
widget in the application).
The widgetinspector still assumes that the overlay is there -> crash.

This patch fixes that by always generating a new overlay widget in case
it is deleted.