Fix regression: Build without graphviz includes
The CMake logic was flawed, in case the libraries were found but the
include headers were not, the script still tried to find out the version
for GraphViz using graphviz_version.h.

This patch checks if graphviz_version.h is available, and only computes
the version *if* it is available.

Allen, please review.