Try harder to get into MSVC builds in release mode.
In release mode MSVC seems to use the counter-part to -Bsymbolic, with so
far no known way to disable that. This means function overwrites don't
actually work. Which is fine, we have an event-filter based fallback that
still sees most objects, but for that we must create the probe late enough,
after qApp exists.