Make compile with both cgraph and graph library
GraphViz offers two libraries to build up graph structures with:
- graph (deprecated)
- cgraph

Prior to this patch we were only supporting the 'graph' approach. This
patch makes it possible to compile both against 'graph' and 'cgraph'.

cgraph seems to be available for some time already, but it is not
stable for graphviz <2.30.x. For me, gammaray instantly crashes when
calling gvContext() in 2.29 (from Ubuntu repositories) when linked
against cgraph. Hence we only use cgraph in case we have >=2.30 and when
cgraph is actually available.

This should also fix the build issues under windows, where the package
provided by graphviz.org only provides the cgraph library.