Remember what the ConnectionFilterProxyModel is filtering on.
This fixes two things:

a) no more columnAdded/Removed signals when the receiver/sender
is set to null when the selection in the scene inspector is removed.
b) again, when the selection in the scene inspector is removed we
would show all connections even though we want to show none in such
a case.