Keep a generic ObjectBroker::object<T>(QString) for the future.
The version which uses the qobject_interface_iid() only works if
just a single object of the given type is ever registered. If multiple
objects for a given interface are registered, we need the generic
function above which this patch re-introduces.

Until the tools are ported, they now use the even more generic
objectInternal function.