Don't crash when remotely invoking a blocking operation.
If that operation creates QObjects, the client gets directly notified
about that and asks for the column count. These requests though will
only be handled once the blocking operation on the server side has
finished. At that point, the objects which where created may be gone
already.

A proper solution to this problem is unclear, i.e. how to still
process client requests even when the server side is blocked.
Threading does not sound like a good idea as it will add other issues.