Remove dependency from the probe on the main window.
This is only needed for filtering UI elements from the object model, so a
simple QObject suffices. In the next step the in-process UI can move to a
separate DLL only loaded on demand.

This brings us back to how to properly communicate settings (like the
search path in this case) to the probe. Environment variables work nicely
in the launch case, but not for attaching.