Prevent flickering/endless loops between NetworkSelectionModels.
If either endpoint changes the selection and the remote end acts
upon it, we must never sent any data back to the initial endpoint.
This would otherwise trigger endless loops/ping-pongs.

It shows btw. that the issue lies in lacking support for extending
of selections, which can be done by selection multiple items in a
list by presseing with the mouse and then hovering over other items.