Properly remote the selectionModelInspector.
This is done by pushing all QObject/QAbstractItemModel and
QItemSelectionModel logic to the server side and publishing a
QIdentityProxyModel which gets the selected source model as
source set. This then magically works on the remote client.

NOTE: This requires we increase the dependency on Qt 4.8. The
alternative would be to copy over KIdentityProxyModel.

What is preferred?