Properly remote the selectionModelInspector.
This is done by pushing all QObject/QAbstractItemModel and
QItemSelectionModel logic to the server side and publishing a
QIdentityProxyModel which gets the selected source model as
source set. This then magically works on the remote client.

In pre-Qt4.7 this uses QSortFilterProxyModel as a fallback
instead of QIdentityProxyModel. It works just like it, just
slower.