Remote the displaying of fatal messages.
This is achieved by forwarding the data to the client and waiting
for the data to finish writing. Then the server exits.

The client gets the message, displays the dialog. The problem
is now that the server disconnected and we currently show another
dialog for that (which hides the backtrace) and on accept closes
the client.