Properly handle RowColumnCountReply for parents without columns/rows.
This will happen e.g. if you connect a client to the server before
any debug message was added to the MessageModel. The reply to the
RowColumnCountRequest returns 0/2. The old code then returns early
and never sets rowCount/columnCount on the parent node to 0/2 but
keeps it at the "loading..." state of -2/-1.

This patch fixes this and we can happily connect the client to the
server and add debug messages later on and they will show up on the
client side. This probably fixes some other cases as well.