Fix assertion in RemoteModel for a thankfully harmless race condition.
The rowCount might be "-1" which would mean we didn't request the
row/col count explicitly. Yet the data is valid and we can use it
nonetheless. This happens in some (thankfully harmless) race
conditions, e.g. when we request row/col count then first get
messages from the server that a row was deleted and that another was
created at the same model index. Then later he'll answer the initial
row/col count request and we handle it here. Since the initial Node
was deleted and a new one created for that index, it's rowCount will
be -1 but the response data is actually useful and we store it.