Fix infinite display of "Loading..." for some columns.
This especially happened for the second column on the top level
of the font model, which returned an empty result set for the
item data. This previously confused our code as we though that
we are still loading in such cases.

Now, instead add the loading message upon request to the data
structure and blindly return what we have in there. When we get
new data, overwrite what we have.

This works reliably and actually looks simpler as well.