Add remote support for the StateMachineViewer.
This mostly boils down to refactoring the StateMachineViewer and
splitting it up into a Server/Client and Widget part.

The interaction between those two parts is handled via the new
Interface which emits signals whenever the graph has changed.

In the future this should probably be rewritten to evaluate the
graph structure on the client side by going through models which
are published from the server side. The Transition and StateModels
might already fit part of the bill and only need to be extended.

To get this working short-term though, the approach taken here should
suffice for now. Problems might arise when connecting a client
to a server which is in the middle of constructing the graph and/or
after it constructed the graph. This only is a problem when we have
multiple clients connected to a single server (or have both, in-probe
and remote client gui).