Be extra sure to not assert in the timertop plugin.
setProperty() must only be called by the owning thread. We already have a
bunch of safeguards for this, but apparently we missed some corner cases,
like QTimers being moved to a different thread.

This at least fixes the symptoms, a proper solution requires refactoring
to use an internal timer to timer info map for example.