First steps towards remoting the SceneInspector.
We introduce an interface and client of the SceneInspector and
create a dummy, empty scene on the client side and so far only
set the proper scene rect on it.

In the future we will then forward interactions with that view
like zooming, rotating etc. to the server and render the scene
and propagate a pixmap of that to the client.

For in-process mode we still use the user scene directly.