Properly setup the layout on the remote client and don't freeze.
We must not set the resize mode on the header of the resource tree
view, as that degrades performance very badly. The reason is that
on a remote client the model is filled in chunks and every time a
new row is inserted it would query the size of _all_ rows again
to resize the column to its contents. This froze the remote client
app when selecting the resource tool.

Now instead we start a timer whenever a row is inserted and once
no rows where inserted for some time we reisze the columns to
their contents. This is fast enough and doesn't freeze the tool on
startup anymore.

The nice side effect is that we finally have the splitter setup
properly on the remote client as well.