Optimize the StyleInspector for remote usage.
We set a fixed cell width and height as defined through the
width/height/zoom controls. These are sent to the remote server
via a shared interface and used in the model's size hint role.

The big advantage is that this way, we don't need to query the
SizeHintRole (which would also query the other data, i.e. the
pixmaps) for the _full_ model content, which is what the
resizeColumnToContent would do.

Overall this makes the StyleInspector much much quicker to
initialize and use from a remote client, as the pixmaps are now
loaded on demand for the visual section only.