Make remote ResourceBrowser work with Qt5.
Apparently, in Qt4 this only worked due to being less strict when
it comes to method invokation. The argument to resourceSelected
was a QVariant which we cannot properly remote. It is our fundamental
building block for serialization after all and thus is always unpacked
on the remote client. Then we invoke the object with the explicit type
the variant contains. Since QVariant has no implicit ctor for QPixmap
it fails in Qt5 and is unclear why it works in Qt4...

This way though the code is clean and works in both versions.