Make it possible to use ObjectVisualizer on servers without VTK.
Only the client/GUI needs VTK, the server-side model works fine
without it.

This is a bit hacky as it will now always show the Object Visualizer
tool and you can select it in clients, which where built without VTK.
Then an error message is shown. The proper way, of having the tool
disabled, won't work until we properly split client/server factories.

This also makes it possible to visualize objects in a Qt5 app, as
long as you use a Qt4 GammaRay client.