Group paint analyzer update requests to prevent flooding.
Previously, if one navigated through the list of the paint buffer,
one would trigger a repaint for every step. Thus, if one navigated
rapidly over a long list, one could easily freeze a low-end embedded
device this way.

Now, we group the repaint requests with a 100ms timer, similar to
how we already do it for the widget paint preview.