Reset the loading state of nodes that moved.
If we have requested data of a cell before a move, the corresponding reply
possibly contains the outdated index, and we then never request the content
for the new location. This fixes cells stuck at "Loading...", especially
with objects that are quickly created and destroyed, such as a QML
ListView.