Add a way to trigger sg visualize overdraw feature
The Qt Quick scene graph has a feature to draw a 3D box visualizing
the drawn items in layers, helping to find overdraw. This commit adds
a button to the gammaray quick-inspector to trigger this feature.

Moreover with this commit the quickinspector is able to tell the
client, what kind of features it supports and the client now hides the
appropriate UI if a feature is not supported (e.g. the Scene Graph tab
when the scene graph inspecting feature is not available).