Add feature to navigate to object in property tab.
This commit adds a feature to navigate to the value of a property in
the property tab, if it's a QObject or a type known to GammaRay's
MetaObject system. The property will be opened inside the appropriate
tool. So for example, if you view an object inside the object
inspector and that object has a property "myStateMachine" which is of
the type QStateMachine, you can right click on the property, choose
"Show in State Machines" and the state machine will be opened there.
The same of course works as well to navigate to the parent object of
some QObject and view it in the same tool.