Visualize item flags in QuickInspector's item tree
With this commit the item tree view in quickinspector visualizes item
flags, which are added in this commit as well. Item give some random,
but useful infos about an item, like is invisible, has active focus,
etc.
These infos are shown in the tooltip as well as by showing icons in
the tree view or changing the text color.
Moreover with this commit items that changed recently are highlighted
for a few seconds using a red text, animated to fade back to black.