QuickItemTree: Skip root item in out-of-view check
In Qt 5.2 the root item of a scene has a size of 0. This caused all
items to be considered to be out of view. Now the root item's (only)
child (the outer-most item, that is explicitly defined in QML), which
has a correct size is treated as the scene size, thus now the check
is working correctly with Qt 5.2, as well.