Quick item tree: No animation on initial load.
This commit removes the animation, where the text in the item tree
view fades from red to black, for the initial load.
Prior to this change on the initial load all items would be red,
fading out together to black. Now the animation happens only when the
data actually changed. This is done using the loading state
information from the remote model.