Refactor the QtQuick scene preview.
The drag feature is now based on a Flickable, instead of implementing
it manually.

Moreover QuickItemOverlay was renamed to AnnotatedScenePreview and now
also takes care of painting the image. This works around the issue of
not being able to update the data from the QQuickImageProvider
properly. The former workaround caused the rulers to completely
recreate themselves on each frame, because the size would change to 0
(-> no ruler at all) and back to the actual size.

In addition to that, the measurement-feature, available with CTRL +
drag, now paints a line that indicates what is actually measured. Also
it's now aligned to full (zoomed) pixels.