QtQuick: Fetch preview using ShaderEffectSource.
The QuickInspector used to use QQuickWindow::grabWindow in order to
fetch the preview image, but that is quite buggy and likes to crash in
obscure ways. Now the inspector inserts a ShaderEffectSource into the
scene and uses it to fetch the data. This is (seems to be) more solid.
Moreover it opens up the possibility to fetch previews of single
items, no matter if they are covered.