Fix deselect of tool in quick scene preview.
When clicking a selected tool in the toolbar of the quick inspector's
scene preview, the render mode would be deactivated and the button's
state would be reset to checked=false, but the ExclusiveButtonGroup
would still have that button set as the current, causing it not to be
able to be reactivated without activating another tool before. Now the
current button is reset to null, causing the button to be unchecked
and the internal state to be correct.