Update margins correctly on size change.
The quick scene preview image has a margin of the size of the preview
panel, in order to scroll half over the image size (in order to have
e.g. the right end of the scene displayed in the center of the panel).
This margin was not correctly updated when the panel size changed.