Quick: Add toolbar buttons for hidden features.
The preview has a measure and a redirect-input features to be
activated using keyboard modifiers. They were pretty hidden, though.
This commit adds toolbar buttons for them, so they gain more
visibility.

Also this commit changes the keyboard-shortcut for redirect-input from
Ctrl+Shift to Shift only, in order to not conflict with select-item,
when we add this feature later. (select-item is Ctrl+Shift in
widget-inspector)