Add possibility to call Q_INVOKABLES.
The methodContextMenu() method used to have a check if the method is a
slot before providing the option to invoke. This commit extends this
check by also providing invoke for Q_INVOKABLEs.

Moreover the method now returns if there is no action to take instead
of trying to open an empty context menu (which resulted in an assert).