Add support for registering multiple signal spy callback sets.
Qt only supports a single one, which was used by the TimerTop plug-in so
far. Instead we now have a list of callback sets in the probe, and a single
registered callback that just dispatches to that list of callbacks.

This allows additional plug-ins to use the same feature, such as Mathias'
signal tracer.