Unify the implementation of the object list and tree models.
Just like the tree model, we now use the infrastructure for multi-thread
support from Probe, rather than rolling our own. This also means we are
now back to a recursive mutex, which fixes a deadlock when requesting
the decoration role on the list model (due to an internal temporary QFile
during icon loading).