Remove the second model hierarchy level, no longer needed.
This massively reduces memory consumption on the client. Obviously memory
growth is still unbound, but at least now at a rate that makes it survive
a few hours rather than a few minutes. If needed we could address this with
suspending/resuming the recording on demand, but it's no longer a blocker
I think.