Avoid using a recursive shader effect source to grab a screenshot.
If there's only one root item (usually that's the case), just position
us next to that and grab its content, non-recursively. If there happen
to be more items already, we keep the current behavior of capturing the
parent content item.

Not using the recusrive mode adds the benefit of not re-rendering when the
content didn't change, and it avoids a crash in the Vivante driver on an
iMX6 system.