Introduce ProbeGuard.
This is a simple RAII class to mark code that is part of the GammaRay
probe. This allows us to check in a few places where we are doing
expensive or otherwise problematic operations inside the probe if we
are dealing with activity from the probe itself (which thus can be
ignored), or activity from the application itself.

This improves performance, obviously, and the quality of the results, as
it allows for example to separate GammaRay debug output from the one of
the target application.