Replace the read/write object lock with a simple mutex.
We have a reader to writer ratio of only about 2:1, so the overhead is
likely not worth it. Also, we can end up in tricky scenarios where we
try to obtain a write lock on an already existing read lock, which will
dead-lock. The unit tests manage to trigger that occasionally, a problem
that goes completely away with a plain mutex.