Rework reparenting handling.
Do not immediately act upon ChildRemoved, but wait if we possibly get a
ChildAdded or ParentChanged (widgets only) event afterwards, as only then
we actually know the move destination. Unfortunately we don't get a
post-move event for moving to a null parent, so we need to process those
after the next event loop re-entry.