Add handler for SIGHUP, SIGINT and SIGTERM.
When receiving a signal to quit in the launcher, we want to quit the
debuggee as well. When hitting CTRL+C this happens, because it will
send SIGINT to the whole process group. But when only gammaray (a.k.a.
the launcher) receives the signal (e.g. killall -2 gammaray) the
debuggee won't quit.
The new handler simply calls QCoreApplication::quit, which causes the
client process to be destroyed.