Let the item tree watcher handle column resizing.
The result is almost as good as ResizeToContent, but at a fraction of the
cost. QHeaderView looks for dataChanged(), which can occur in massive
storms, while the item tree watcher deals with row insertions, and more
importantly, batches thereof.

This removes the major bottleneck in the client when dealing with item
flag change storms.