Give grabWindow another chance.
In the quick inspector initially we used grabWindow to get the preview
of the scene. Later, because grabWindow was crashy, we inserted a
QQuickShaderEffectSource into the scene to get the preview. But
1) this one is crashy too
2) we grabbed the window from the main thread, where we couldn't be
   sure, that the qopengl context is valid (which caused the crashes)
Now, grabbing from the render thread directly after swapping frames,
we retry to use grabWindow, because it's much easier and seems to
nowadays work reliably.