Avoid emitting objectCreated/objectDestroyed signals in the wrong order.
This could happen is objects created/destroyed in a secondary thread in
rapid succession got the same memory address again. To deal with this
we need to ensure we have full control over the emission order, by using
a single queue for them, and by dealing with the thread handover ourselves
rather than relying on queued connections in a few places.