Don't mess with export macros.
This wasn't a problem so far, but with the move into the library it leaks
into the joint moc file, and causes linker errors on Windows.

As there is no longer an upstream copy of this to keep in sync, it also
doesn't matter anymore to minimize the diff, so just remove all traces of
export macros here.