Always assume reliable object tracking when we have qHooks.
This fixes bizarre shutdown crashes with multi-threaded targets due to
the probe connecting to the destroyed() signal of short-lived objects
whose dtor has begun executing already. This results in the connection
being added after the dtor disconnected all connections, and thus the
connection staying alive even after the sender has been destroyed. That
in turn backfired when the receiver (GammaRay::Probe) tried to clean up
the connections on its destruction.