Fix exponential property adaptor leak when reloading sub-trees.
While reloading we can get rowCount() calls as a result to the change
notifications we send out, and trigger the on-demand adaptor creation
there. We then create another adaptor as part of the sub-tree reloading,
and lose the reference to the other one. It's signals are still connected
however, so that with every reload we double the amount of change signals
we get, and thus the amount of sub-tree reloads we do.

The fix is somewhat ugly, and probably should be replaced with a more
general sub-tree blocking solution to optimize loop prevention etc. as
well, but it gets the job done for now.