Fix server-side graph building
We need to always notify the client about the parent of a node in the
graph first. Make the graph building more efficient by doing a BFS
instead of a DFS in the QObject-tree. This also fixes an issue where
parent nodes were left out in the graph building process.

Fixes the state machine for the pinball.scxml example

Task-Id: KDESME-12