Add infrastructure for plugins to provide basic object information.
This covers e.g. names and sources locations, and is needed as we have
more tools that can show QML-created objects. So we need to have the
extra QML object information available generically, not just in the QQ2
inspector.

In a first step this just covers names, making QML ids visible in all
object models in all tools.