Batch dataChanged() signals for content loading.
This is optimized for a minimal amount of dataChanged() signals, rather
than minimal coverage of the changed area, in a first step. This allows
us to use a very simple group by parent/bounding rect approach, which
seems to improve performance considerably on models with columns resized
to content (as each dataChanged causes expensive text layout computations
there).