Give the injectors explicit control over resending the server address.
This is important as some of the less elegant injectors (e.g. style) might
end up with multiple calls to gammaray_probe_inject in rapid succession,
the second one of those might or might not find a launcher to talk to
anymore and then get stuck there for the local socket connection timeout.