Remember the last explicitly selected property tab, and restore it.
This makes the property tab selection behave much more like what one
would expect when working with a specific tool and switching through
objects that might or might not support it. Unless you explicitly
switch to a different tab, it'll now try to restore the selection even
if a tab was temporarily not available.