Implement code navigation for QObjects.
Allows for QObjects to navigate to the code location, where they were
created. Not (only) the header file or the class's constructor, but
where the constructor was called from.

This is done by generating a backtrace from
Probe::objectAdded, which is called from the QObject constructor. From
there we can step up to the frame above the object's constructor and
get the source location of that frame.

The stacktrace generation is based on backward-cpp.