Free tool handling from models.
This change removes ToolModel and ToolSelectionModel. It replaces them
by a new ToolManager class, that does the same. On the client side I
added the class ClientToolManager to provide the new tool management
API. It also offers it's data as a model, but the new model is client
sided only.

Now you can also access the tools directly without
the model. That is needed if we want to explicitly show a specific
tool instead of just displaying a list of all available tools, e.g.
when integrating into an IDE.

Resolves #229