Distinguish between gadget values and gadget pointers
Gadget values need to be treated like meta object values, ie. we need
to propagate writes up the hierarchy, and prevent writes if the path to
the top is blocked somewhere by a read-only element.

This also expands the unit tests to cover value type write propagation
already, which yet has to be implemented.

This fixes e.g. QSizePolicy property trees being editable although we
can't propagate the changes yet.