Fix Qt Core and ABI detection on embedded Linux systems without ldd
This patch adds a fallback which gets run when ldd could not be
started. In that case, we instead launch the target debuggee binary
directly, but set the LD_TRACE_LOADED_OBJECTS=1 environment
variable. This makes ld.so behave like ldd, and instead of executing
the debuggee it will output the desired information about dynamic
dependencies.