Add QuickOverlay for quick inspector
This mimic the bahavior of the widget inspector
The client is still drawing its decoration client side to avoid
scaled rendering which make it unreadable when using the client zoom.
The server is also drawing its decorations but they do not appear in
sent frame to the client to avoid the double decorations drawing.
This ensure that both the client and server have their decorations
independently of each other.
We deprecated usage of grabWindow/grabFrameBuffer.
This allow to grab content just before drawing decoration and send it
to the client avoiding flickering.
This also avoid costly grabWindow/grabFrameBuffer.

Task-Id: KDEND-108