QtIvi plugin: Implement object navigation with other plugins.
This allows to navigate from / to QtIvi to / from other plugins
such as QObject, when invoking the context menu on a suitable
object. Unfortunately it doesn't work for parents of QtIviProperty
instances because the parents are not of a special class. The
navigation is based on class names in GammaRay common code.

Task-Id: KDEND-85