Meta Objects: Don't warn about internal methods
With the introduction of private _q_createJSWrapper() in QQuickItem the
entire class hierarchy gets flagged in the meta object browser for having
slots that aren't dynamically invokable. The warning is of dubious use
for private methods though, as invoking those dynamically is far less
common/needed.

So let's just warn about issues in the public interface.