State Machines: Fix selection model syncing
QtSCXML does not expose pointers to instances for states/transitions, thus we need a
different way to address the elements of the state machine. There's
already the internal id for states/transitions, exposed as
GammaRay::{StateId,TransitionId}.

Use this instead of depending on ObjectId which is only available for objects known to
GammaRay.

Note: One problem is that KDSME uses quintptr as internal ids for the
state chart elements, while GammaRay uses quint64. So in the unlikely
case that the development host is not 64 bit and the target *is* 64 bit,
we might not be able to sync the selection models correctly. To fix we would need
changes in KDSME though.

Task-Id: KDEND-127