QuickInspector: Correctly track changing windows
The window of an item changes much more often, than expected before.
Changes in window of the item are tracked over the whole item's lifetime such
that it gets properly re-added to a newly started quickitem-scene.
This prevents missing items when inspecting a scene which is still in
its upstart phase.