WIP Add support for finding dependencies.
Including loop detection. Finding dependencies works only for
dependencies to scope and context properties, though, not for
dependencies to id objects. This is hopefully going to come later.

The code is still very much work in progress.

This commit also contains code to find the expression code of a binding.
This currently works by simply opening the qml file and reading the
corresponding line of code. Not very satisfying, though.