Don't create new TransformNodes triggered by Gammaray
This fixes a SEGFAULT where in some cases an invalid Transform Node was
implicitly created through QQuickItemPrivate::itemNode.
In the case that no ItemNodeInstance is created, we just skip this item.

Since this is only relevant in individual frames, it does not exclude
the formerly skipped items, as they are re-added in a frame where they
are actually initialized by the scenegraph.

Closes #327