Add generic Yocto Toolchain file
Use this toolchain file when a Yocto environment setup file is present.
This is the case when a yocto environment was used for the toolchain build
image and sysroot from tqtc's online installer in Qt for Device Creation
as well as Qt Automotive suite also have this.

Reduces the required cmake call when crosscompiling to:
cmake -DCMAKE_TOOLCHAIN_FILE=<path/to/src>/cmake/Toolchain-Boot2Qt.cmake -DCMAKE_INSTALL_PREFIX=<install> <path/to/src>